Move more element related components into src/components/elements
