+++ app/packs/src/components/elements/ElementDetails.js
